refactor(fhir): pull up common VS ops logic in FhirValueSetRequest
